Update T1037: Atomics "Logon Scripts" and "Startup Folder Script" (#762)
* The Atomic tests "Logon Scripts" and Startup Folder Script" were updated
with additional input arguments. The first test required a fix to the
string type for the registry entry to allow it to function correctly.
Added a log file write command for each test to record if the commands ran at startup
correctly. Other minor syntax and description updates.

* Added cleanup commands to cleanup new run-log files added to verify success
of test.

Co-authored-by: Carrie Roberts <clr2of8@gmail.com>