atomic added to T1021.006 (#1116)
* T1021.006 evil-winrm atomic

* Update T1021.006.yaml

* Update T1021.006.yaml

fixed input args

* Update T1021.006.yaml

added Prereqs for Ruby and moved Evil-WinRM to a Prereq

* Update T1021.006.yaml

removed duplicate description and changed Ctrl + C to exit.

* Updated yaml

updated descriptions for prereqs. removed un-needed "exit" from cleanup_command.

* $env:username replaced

$env:username replaced with $env:Temp to account for when people have who have user profiles in althernative locations than C and also download to TEMP instead of Desktop.

* Removing cleanup_command

Removing cleanup_command as the evil-winrm is a prereq gem. in the future, if a cleanup_prereq_command is implemented this may be worth adding back in(gem uninstall evil-winrm -x).

Co-authored-by: Carrie Roberts <clr2of8@gmail.com>