Fix Golden Ticket that seemed to work even with wrong parameters
Before the golden ticket was injected in the current session which had the issue, first, of clearing the current tickets in the current session. Then, it seemed to work with an invalid Golden Ticket because somehow Windows managed to get the valid normal TGT and use it instead.
Now we create a new empty windows session (with runas /netonly) and work in it so we don't pollute the current normal session, and we don't have false positives.

Also, remove the "machine" parameter since we now use the domain SYSVOL share to trigger a TGS request. So one less dependency (needed an additional target machine in domain)