Update T1098.yaml
One of the steps the Adversary might perform so it can make other changes to the system in order to maintain persistance is Activating a Guest account.  This account is deactivated in most of the windows operating systems. So, all adversary needs to do is to activate this account by using a simple net user command. This is a part of the overal account manipulation Technique under https://attack.mitre.org/techniques/T1098/.