Added docker testcase for T1609.yaml
Fairly straight forward way to execute into a container.

Details:
The test was created for kubernetes and mitre framework also mentioned about docker. So created a second test for the same.

Testing:
Tested using ubuntuas the base image. If using just Docker, run the container(already present in the script) and execute the testcase.

It may take up to a minute to run due to image creation. If it hangs for longer than a minute, something went wrong.

Associated Issues:
None