Added Deploying a docker
Fairly straight forward to deploy a container.

Details:
There was no test case for deploying a container in the atomic-red team and I was working with atomic red team so thought to create a pr for this feature

Testing:
Tested using ubuntu as the base image. It creates an image and runs that image to check that image is deployed properly.

It may take up to a couple of minutes to run due to image creation. If it hangs for longer than a minute, something went wrong.

Associated Issues:
None