Create T1027.011
Demonstrates storing and retrieving data filelessly using the Windows Registry. The test first checks if the specified registry key path exists; if not, it creates the path. It then generates a simple test value, stores this data in the registry, and retrieves it. The registry key is removed in the cleanup phase.