Update T1560.001.yaml
This PowerShell script illustrates a popular data exfiltration method as part of my research. After copying all of the data from the user's AppData folder to the desktop and compressing it into a ZIP file for easy use, it deletes the temporary copy to prevent any traces. To function effectively, the script needs administrative privileges and has error handling for folders that are locked or inaccessible.

For both offensive testing and defensive awareness, it is helpful to comprehend this technique since it resembles the strategies used by attackers to surreptitiously gather important user data.

GitHub: HyperionRising12.